Change environment exception to warning
If `Envgetter` is unable to provide a "complete environment",
apparently based on environmental variables, an exception is raised
and testing stops. However, most of these issues arise from deep
diving into bash functions and other oddities. The vars that the
script are worried about are normal vars and will be replicable.

Hence, changing the `Exception` to a `warnings.warn` so the user is
aware of the situation and be able to, if necessary, debug later
issues in a test. However, the testing will at least proceed.